Undo fusion stream changes to common
