+++ crates/burn-fusion/src/stream/mod.rs
@@ -8,5 +8,3 @@ mod multi;
 pub use base::*;
 pub use context::*;
 pub use multi::*;
-
-pub use burn_common::stream::StreamId;
